+++ service/src/main/kotlin/io/provenance/explorer/domain/models/explorer/ExplorerModels.kt
@@ -18,7 +18,7 @@ data class BlockSummary(
 data class Spotlight(
     val latestBlock: BlockSummary,
     val avgBlockTime: BigDecimal,
-    val bondedTokens: BondedTokens,
+    val bondedTokens: CountStrTotal,
     val totalTxCount: BigInteger
 )